requirements.txt
asgiref==3.2.7
Babel==2.8.0
cffi==1.14.0
cryptography==2.9
Django==3.0.5
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
phonenumbers==8.12.1
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
